Enable link during portable publish (dotnet/linker#263)
Commit migrated from https://github.com/dotnet/linker/commit/16d3016328b310f035e50b1f376e054f45c4e5dd

+++ src/tools/illink/corebuild/integration/test/WebApiTest.cs
@@ -8,11 +8,21 @@ namespace ILLink.Tests
 {
 	public class WebApiTest : IntegrationTestBase
 	{
-		public WebApiTest(ITestOutputHelper output) : base(output) {}
+		private string csproj;
+
+		public WebApiTest(ITestOutputHelper output) : base(output) {
+			csproj = SetupProject();
+		}
 
 		public string SetupProject()
 		{
 			string projectRoot = "webapi";
+			string csproj = Path.Combine(projectRoot, $"{projectRoot}.csproj");
+
+			if (File.Exists(csproj)) {
+				output.WriteLine($"using existing project {csproj}");
+				return csproj;
+			}
 
 			if (Directory.Exists(projectRoot)) {
 				Directory.Delete(projectRoot, true);
@@ -25,7 +35,10 @@ namespace ILLink.Tests
 				Assert.True(false);
 			}
 
-			string csproj = Path.Combine(projectRoot, $"{projectRoot}.csproj");
+			PreventPublishFiltering(csproj);
+
+			AddLinkerReference(csproj);
+
 			return csproj;
 		}
 
@@ -49,18 +62,23 @@ namespace ILLink.Tests
 		}
 
 		[Fact]
-		public void RunWebApi()
+		public void RunWebApiStandalone()
 		{
-			string csproj = SetupProject();
+			string executablePath = BuildAndLink(csproj, selfContained: true);
+			CheckOutput(executablePath, selfContained: true);
+		}
 
-			PreventPublishFiltering(csproj);
+		[Fact]
+		public void RunWebApiPortable()
+		{
+			string target = BuildAndLink(csproj, selfContained: false);
+			CheckOutput(target, selfContained: false);
+		}
 
-			AddLinkerReference(csproj);
-
-			BuildAndLink(csproj);
-
+		void CheckOutput(string target, bool selfContained = false)
+		{
 			string terminatingOutput = "Now listening on: http://localhost:5000";
-			int ret = RunApp(csproj, out string commandOutput, 60000, terminatingOutput);
+			int ret = RunApp(target, out string commandOutput, 60000, terminatingOutput, selfContained: selfContained);
 			Assert.True(commandOutput.Contains("Application started. Press Ctrl+C to shut down."));
 			Assert.True(commandOutput.Contains(terminatingOutput));
 		}
